Bug fix. Requires clearing db and re-running rake.

+++ app/models/sales_aggregate.rb
@@ -3,7 +3,7 @@ class SalesAggregate
   
     field :tag, type: String
     field :qty, type: Integer
-    field :value, type: BigDecimal
+    field :value, type: Float
     
     belongs_to :made_in, class_name: "Locale", inverse_of: :sells
     belongs_to :sold_in, class_name: "Locale", inverse_of: :buys
